+++ app/routes/components/InputBase/index.tsx
@@ -1,12 +1,10 @@
-import type { Route } from "./+types/index";
-
 import { ComponentTemplate } from "src/components/ComponentTemplate";
 import { Section } from "src/components/Section";
-
-import { sourceCode } from "./source_code.codegen";
+import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
+import { sourceCode } from "./source_code.codegen";
 
-export function meta({}: Route.MetaArgs) {
+export function meta(_: Route.MetaArgs) {
   return [
     { title: "InputBase | Copy UI" },
     { name: "description", content: "The InputBase component from Copy UI." },
@@ -24,4 +22,3 @@ export default function InputBasePage() {
     </ComponentTemplate>
   );
 }
-