+++ src/components/InputBase/index.tsx
@@ -1,3 +1,3 @@
-// InputBase from copy-ui @ 2025-09-30.
+// InputBase from copy-ui @ 2025-10-02.
 
 export { InputBase } from "./InputBase";